move the pdf creation to generatePdf in utils folder and import it in MiroSelectionBox
